pyproject.toml
[tool.poetry]
name = "learn-drf"
version = "0.1.0"
description = ""
authors = ["pythonuzbek <nusratullo3304@gmail.com>"]
readme = "README.md"
packages = [{include = "learn_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
django-rest-framework = "^0.1.0"
drf-yasg = "^1.21.5"
django-filter = "^23.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
